move everything into separate bitarray
